+++ core/src/main/java/org/jdbi/v3/core/StatementContext.java
@@ -13,19 +13,16 @@
  */
 package org.jdbi.v3.core;
 
-import static java.util.Objects.requireNonNull;
+import org.jdbi.v3.core.argument.Argument;
+import org.jdbi.v3.core.mapper.ColumnMapper;
+import org.jdbi.v3.core.mapper.RowMapper;
 
 import java.lang.reflect.Type;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.util.Arrays;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collector;
-
-import org.jdbi.v3.core.argument.Argument;
-import org.jdbi.v3.core.mapper.ColumnMapper;
-import org.jdbi.v3.core.mapper.RowMapper;
 
 /**
  * The statement context provides a means for passing client specific information through the
@@ -33,37 +30,7 @@ import org.jdbi.v3.core.mapper.RowMapper;
  * to all statement customizers. This makes it possible to parameterize the processing of
  * the tweakable parts of the statement processing cycle.
  */
-public final class StatementContext
-{
-    private final JdbiConfig config;
-    private final ExtensionMethod extensionMethod;
-
-    private final Cleanables cleanables = new Cleanables();
-
-    private String            rawSql;
-    private String            rewrittenSql;
-    private PreparedStatement statement;
-    private Connection        connection;
-    private Binding           binding;
-    private boolean           returningGeneratedKeys;
-    private boolean           concurrentUpdatable;
-    private String[]          generatedKeysColumnNames;
-
-    public StatementContext() {
-        this(new JdbiConfig());
-    }
-
-    StatementContext(JdbiConfig config)
-    {
-        this(config, null);
-    }
-
-    StatementContext(JdbiConfig config, ExtensionMethod extensionMethod)
-    {
-        this.config = requireNonNull(config);
-        this.extensionMethod = extensionMethod;
-    }
-
+public interface StatementContext {
     /**
      * Specify an attribute on the statement context
      *
@@ -72,10 +39,7 @@ public final class StatementContext
      *
      * @return previous value of this attribute
      */
-    public Object setAttribute(String key, Object value)
-    {
-        return config.statementAttributes.put(key, value);
-    }
+    Object setAttribute(String key, Object value);
 
     /**
      * Obtain the value of an attribute
@@ -84,10 +48,7 @@ public final class StatementContext
      *
      * @return the value of the attribute
      */
-    public Object getAttribute(String key)
-    {
-        return config.statementAttributes.get(key);
-    }
+    Object getAttribute(String key);
 
     /**
      * Obtain all the attributes associated with this context as a map. Changes to the map
@@ -95,10 +56,7 @@ public final class StatementContext
      *
      * @return a map f attributes
      */
-    public Map<String, Object> getAttributes()
-    {
-        return config.statementAttributes;
-    }
+    Map<String, Object> getAttributes();
 
     /**
      * Obtain a column mapper for the given type in this context.
@@ -106,10 +64,7 @@ public final class StatementContext
      * @param type the target type to map to
      * @return a ColumnMapper for the given type, or empty if no column mapper is registered for the given type.
      */
-    public Optional<ColumnMapper<?>> findColumnMapperFor(Type type)
-    {
-        return config.mappingRegistry.findColumnMapperFor(type, this);
-    }
+    Optional<ColumnMapper<?>> findColumnMapperFor(Type type);
 
     /**
      * Obtain a row mapper for the given type in this context.
@@ -117,10 +72,7 @@ public final class StatementContext
      * @param type the target type to map to
      * @return a RowMapper for the given type, or empty if no row mapper is registered for the given type.
      */
-    public Optional<RowMapper<?>> findRowMapperFor(Type type)
-    {
-        return config.mappingRegistry.findRowMapperFor(type, this);
-    }
+    Optional<RowMapper<?>> findRowMapperFor(Type type);
 
     /**
      * Obtain an argument for given value in this context
@@ -128,46 +80,27 @@ public final class StatementContext
      * @param value the argument value.
      * @return an Argument for the given value.
      */
-    public Optional<Argument> findArgumentFor(Type type, Object value) {
-        return config.argumentRegistry.findArgumentFor(type, value, this);
-    }
+    Optional<Argument> findArgumentFor(Type type, Object value);
 
     /**
      * Obtain a collector for the given type in this context
      * @param type the result type of the collector
      * @return a Collector for the given result type, or null if no collector factory is registered for this type.
      */
-    public Optional<Collector<?, ?, ?>> findCollectorFor(Type type) {
-        return config.collectorRegistry.findCollectorFor(type);
-    }
+    Optional<Collector<?, ?, ?>> findCollectorFor(Type type);
 
     /**
      * @param containerType the container type.
      * @return the element type for the given container type, if available.
      */
-    public Optional<Type> elementTypeFor(Type containerType) {
-        return config.collectorRegistry.elementTypeFor(containerType);
-    }
-
-    void setRawSql(String rawSql)
-    {
-        this.rawSql = rawSql;
-    }
+    Optional<Type> elementTypeFor(Type containerType);
 
     /**
      * Obtain the initial sql for the statement used to create the statement
      *
      * @return the initial sql
      */
-    public String getRawSql()
-    {
-        return rawSql;
-    }
-
-    void setRewrittenSql(String rewrittenSql)
-    {
-        this.rewrittenSql = rewrittenSql;
-    }
+    String getRawSql();
 
     /**
      * Obtain the located and rewritten sql
@@ -177,15 +110,7 @@ public final class StatementContext
      *
      * @return the sql as it will be executed against the database
      */
-    public String getRewrittenSql()
-    {
-        return rewrittenSql;
-    }
-
-    void setStatement(PreparedStatement stmt)
-    {
-        statement = stmt;
-    }
+    String getRewrittenSql();
 
     /**
      * Obtain the actual prepared statement being used.
@@ -195,74 +120,25 @@ public final class StatementContext
      *
      * @return Obtain the actual prepared statement being used.
      */
-    public PreparedStatement getStatement()
-    {
-        return statement;
-    }
-
-    void setConnection(Connection connection)
-    {
-        this.connection = connection;
-    }
+    PreparedStatement getStatement();
 
     /**
      * Obtain the JDBC connection being used for this statement
      *
      * @return the JDBC connection
      */
-    public Connection getConnection()
-    {
-        return connection;
-    }
+    Connection getConnection();
 
-    void setBinding(Binding b)
-    {
-        this.binding = b;
-    }
+    Binding getBinding();
 
-    public Binding getBinding()
-    {
-        return binding;
-    }
-
-    Cleanables getCleanables() {
-        return cleanables;
-    }
-
-    public ExtensionMethod getExtensionMethod()
-    {
-        return extensionMethod;
-    }
-
-    void setReturningGeneratedKeys(boolean b)
-    {
-        if (isConcurrentUpdatable() && b) {
-            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
-                    + "updatable and is returning generated keys.");
-        }
-        this.returningGeneratedKeys = b;
-    }
+    ExtensionMethod getExtensionMethod();
 
     /**
      * @return whether the statement being generated is expected to return generated keys.
      */
-    public boolean isReturningGeneratedKeys()
-    {
-        return returningGeneratedKeys || generatedKeysColumnNames != null && generatedKeysColumnNames.length > 0;
-    }
+    boolean isReturningGeneratedKeys();
 
-    public String[] getGeneratedKeysColumnNames()
-    {
-        if (generatedKeysColumnNames == null) {
-            return new String[0];
-        }
-        return Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
-    }
-
-    void setGeneratedKeysColumnNames(String[] generatedKeysColumnNames)
-    {
-        this.generatedKeysColumnNames = Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
-    }
+    String[] getGeneratedKeysColumnNames();
 
     /**
      * Return if the statement should be concurrent updatable.
@@ -273,24 +149,5 @@ public final class StatementContext
      *
      * @return if the statement generated should be concurrent updatable.
      */
-    public boolean isConcurrentUpdatable() {
-        return concurrentUpdatable;
-    }
-
-    /**
-     * Set the context to create a concurrent updatable result set.
-     *
-     * This cannot be combined with {@link #isReturningGeneratedKeys()}, only
-     * one option may be selected. It does not make sense to combine these either, as one
-     * applies to queries, and the other applies to updates.
-     *
-     * @param concurrentUpdatable if the result set should be concurrent updatable.
-     */
-    void setConcurrentUpdatable(final boolean concurrentUpdatable) {
-        if (concurrentUpdatable && isReturningGeneratedKeys()) {
-            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
-                    + "updatable and is returning generated keys.");
-        }
-        this.concurrentUpdatable = concurrentUpdatable;
-    }
+    boolean isConcurrentUpdatable();
 }
