+++ core/src/test/java/org/jdbi/v3/core/JdbiAccess.java
@@ -34,6 +34,6 @@ public class JdbiAccess {
      * with the given handle.
      */
     public static StatementContext createContext(Handle handle) {
-        return new StatementContext(handle.config);
+        return new ConcreteStatementContext(handle.config);
     }
 }
